+++ experimental/packages/opentelemetry-sdk-node/src/sdk.ts
@@ -113,13 +113,13 @@ function getValueInMillis(envName: string, defaultValue: number): number {
  */
 function configureMetricProviderFromEnv(): IMetricReader[] {
   const metricReaders: IMetricReader[] = [];
-  const enabledExporters = getStringListFromEnv('OTEL_METRICS_EXPORTER');
-  if (!enabledExporters) {
-    return metricReaders;
-  }
+  const enabledExporters = Array.from(
+    new Set(getStringListFromEnv('OTEL_METRICS_EXPORTER') ?? [])
+  );
 
   if (enabledExporters.length === 0) {
     diag.debug('OTEL_METRICS_EXPORTER is empty. Using default otlp exporter.');
+    enabledExporters.push('otlp');
   }
 
   if (enabledExporters.includes('none')) {
@@ -132,8 +132,10 @@ function configureMetricProviderFromEnv(): IMetricReader[] {
   enabledExporters.forEach(exporter => {
     if (exporter === 'otlp') {
       const protocol =
-        process.env.OTEL_EXPORTER_OTLP_METRICS_PROTOCOL?.trim() ||
-        process.env.OTEL_EXPORTER_OTLP_PROTOCOL?.trim();
+        (
+          getStringFromEnv('OTEL_EXPORTER_OTLP_METRICS_PROTOCOL') ??
+          getStringFromEnv('OTEL_EXPORTER_OTLP_PROTOCOL')
+        )?.trim() || 'http/protobuf'; // Using || to also fall back on empty string
 
       const exportIntervalMillis = getValueInMillis(
         'OTEL_METRIC_EXPORT_INTERVAL',
@@ -461,7 +463,9 @@ export class NodeSDK {
   }
 
   private configureLoggerProviderFromEnv(): void {
-    const enabledExporters = getStringListFromEnv('OTEL_LOGS_EXPORTER') ?? [];
+    const enabledExporters = Array.from(
+      new Set(getStringListFromEnv('OTEL_LOGS_EXPORTER') ?? [])
+    );
 
     if (enabledExporters.length === 0) {
       diag.debug('OTEL_LOGS_EXPORTER is empty. Using default otlp exporter.');
@@ -479,10 +483,11 @@ export class NodeSDK {
 
     enabledExporters.forEach(exporter => {
       if (exporter === 'otlp') {
-        const protocol = (
-          getStringFromEnv('OTEL_EXPORTER_OTLP_LOGS_PROTOCOL') ??
-          getStringFromEnv('OTEL_EXPORTER_OTLP_PROTOCOL')
-        )?.trim();
+        const protocol =
+          (
+            getStringFromEnv('OTEL_EXPORTER_OTLP_LOGS_PROTOCOL') ??
+            getStringFromEnv('OTEL_EXPORTER_OTLP_PROTOCOL')
+          )?.trim() || 'http/protobuf'; // Using || to also fall back on empty string
 
         switch (protocol) {
           case 'grpc':
@@ -492,10 +497,6 @@ export class NodeSDK {
             exporters.push(new OTLPHttpLogExporter());
             break;
           case 'http/protobuf':
-            exporters.push(new OTLPProtoLogExporter());
-            break;
-          case undefined:
-          case '':
             exporters.push(new OTLPProtoLogExporter());
             break;
           default: